Fix #523
Root folder is present in the db with 2 records with the
same remote_id: one is the proper root, the other is the
.Trash folder. When selecting by remote_id, if the trash
folder is returned, it can cause problems to change
detection. So select_resources_with_remote_id is introduced,
and select_resource_with_remote_id is renamed
select_first_resource_with_remote_id.